Create #1

+++ Hotel1/src/com/senlainc/git_courses/java_training/petushok_valiantsin/impl/models/room/Room.java
@@ -3,12 +3,12 @@ package com.senlainc.git_courses.java_training.petushok_valiantsin.impl.models.r
 import com.senlainc.git_courses.java_training.petushok_valiantsin.impl.models.status.Status;
 
 public class Room {
-    private int number;
+    private short number;
     private String classification;
     private int roomNumber;
     private Status status;
 
-    public Room(int number, String classification, int roomNumber, Status status) {
+    public Room(short number, String classification, int roomNumber, Status status) {
         this.number = number;
         this.classification = classification;
         this.roomNumber = roomNumber;
